+++ r_exec/factory.cpp
@@ -36,7 +36,7 @@ namespace	r_exec{
 
 		MkNew::MkNew(r_code::Mem	*m,Code	*object):LObject(m){
 
-			uint32	write_index=0;
+			uint16	write_index=0;
 			code(write_index++)=r_code::Atom::Object(Opcodes::MkNew,5);
 			code(write_index++)=r_code::Atom::RPointer(0);	//	object.
 			code(write_index++)=r_code::Atom::View();
@@ -50,7 +50,7 @@ namespace	r_exec{
 
 		MkLowRes::MkLowRes(r_code::Mem	*m,Code	*object):LObject(m){
 
-			uint32	write_index=0;
+			uint16	write_index=0;
 			code(write_index++)=r_code::Atom::Object(Opcodes::MkLowRes,5);
 			code(write_index++)=r_code::Atom::RPointer(0);	//	object.
 			code(write_index++)=r_code::Atom::View();
@@ -64,7 +64,7 @@ namespace	r_exec{
 
 		MkLowSln::MkLowSln(r_code::Mem	*m,Code	*object):LObject(m){
 
-			uint32	write_index=0;
+			uint16	write_index=0;
 			code(write_index++)=r_code::Atom::Object(Opcodes::MkLowSln,5);
 			code(write_index++)=r_code::Atom::RPointer(0);	//	object.
 			code(write_index++)=r_code::Atom::View();
@@ -78,7 +78,7 @@ namespace	r_exec{
 
 		MkHighSln::MkHighSln(r_code::Mem	*m,Code	*object):LObject(m){
 
-			uint32	write_index=0;
+			uint16	write_index=0;
 			code(write_index++)=r_code::Atom::Object(Opcodes::MkHighSln,5);
 			code(write_index++)=r_code::Atom::RPointer(0);	//	object.
 			code(write_index++)=r_code::Atom::View();
@@ -92,7 +92,7 @@ namespace	r_exec{
 
 		MkLowAct::MkLowAct(r_code::Mem	*m,Code	*object):LObject(m){
 
-			uint32	write_index=0;
+			uint16	write_index=0;
 			code(write_index++)=r_code::Atom::Object(Opcodes::MkLowAct,5);
 			code(write_index++)=r_code::Atom::RPointer(0);	//	object.
 			code(write_index++)=r_code::Atom::View();
@@ -106,7 +106,7 @@ namespace	r_exec{
 
 		MkHighAct::MkHighAct(r_code::Mem	*m,Code	*object):LObject(m){
 
-			uint32	write_index=0;
+			uint16	write_index=0;
 			code(write_index++)=r_code::Atom::Object(Opcodes::MkHighAct,5);
 			code(write_index++)=r_code::Atom::RPointer(0);	//	object.
 			code(write_index++)=r_code::Atom::View();
@@ -120,7 +120,7 @@ namespace	r_exec{
 
 		MkSlnChg::MkSlnChg(r_code::Mem	*m,Code	*object,float32	value):LObject(m){
 
-			uint32	write_index=0;
+			uint16	write_index=0;
 			code(write_index++)=r_code::Atom::Object(Opcodes::MkSlnChg,6);
 			code(write_index++)=r_code::Atom::RPointer(0);	//	object.
 			code(write_index++)=r_code::Atom::Float(value);	//	change.
@@ -135,7 +135,7 @@ namespace	r_exec{
 
 		MkActChg::MkActChg(r_code::Mem	*m,Code	*object,float32	value):LObject(m){
 
-			uint32	write_index=0;
+			uint16	write_index=0;
 			code(write_index++)=r_code::Atom::Object(Opcodes::MkActChg,6);
 			code(write_index++)=r_code::Atom::RPointer(0);	//	object.
 			code(write_index++)=r_code::Atom::Float(value);	//	change.
